bed generation automated; refactor; add tests; add cut_position information;

+++ pavooc/preprocessing/prepare_flashfry.py
@@ -13,15 +13,17 @@ def main():
     except FileExistsError:
         pass
     logging.info('Running FlashFry to build up off-target database')
-    subprocess.Popen([
+    result = subprocess.run([
             'java',
             '-Xmx{}g'.format(JAVA_RAM),
-            '-jar', 'FlashFry-assembly-1.6.jar',
+            '-jar', 'FlashFry-assembly-1.7.jar',
             '--analysis', 'index',
             '--tmpLocation', FLASHFRY_TMP_DIR,
             '--database', FLASHFRY_DB_FILE,
             '--reference', GENOME_FILE,
-            '--enzyme', 'spcas9ngg']).wait()
+            '--enzyme', 'spcas9ngg'])
+    if result.returncode != 0:
+        raise RuntimeError(result)
 
     # free memory
     shutil.rmtree(FLASHFRY_TMP_DIR)
